get periodeTravaillee get periodeTravaillees

+++ src/main/java/LPY/appliVisiteur/Model/Repository/PeriodeTravailleeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.Optional;
 
 public interface PeriodeTravailleeRepository extends CrudRepository<PeriodeTravaillee, Integer> {
-    Optional<PeriodeTravaillee> findById(Integer id);
+    Optional<PeriodeTravaillee> findById(Long id);
 }
